BIG FIX

+++ src/main/java/diploma/dao/UserActivityDao.java
@@ -3,6 +3,6 @@ package diploma.dao;
 import diploma.model.UserActivityEntity;
 
 public interface UserActivityDao {
-    Integer connected(UserActivityEntity userActivityEntity);
-    void disconnected(Integer id, String disconnectionTime);
+    long connected(UserActivityEntity userActivityEntity);
+    void disconnected(long id, String disconnectionTime);
 }
