+++ src/main/java/diploma/dao/UserDao.java
@@ -1,5 +1,5 @@
 package diploma.dao;
 
 public interface UserDao {
-    Integer getId(String login);
+    long getId(String login);
 }
