+++ src/main/java/diploma/services/UserActivityService.java
@@ -1,6 +1,6 @@
 package diploma.services;
 
 public interface UserActivityService {
-    Integer connected(String login, String connectionTime);
-    void disconnected(Integer id, String disconnectionTime);
+    long connected(String login, String connectionTime);
+    void disconnected(long id, String disconnectionTime);
 }
