+++ src/main/java/diploma/services/UserService.java
@@ -1,5 +1,5 @@
 package diploma.services;
 
 public interface UserService {
-    Integer getId(String login);
+    long getId(String login);
 }
